Changes to the install.sh shell script
Hello there!
I rewrote the install.sh file with sepated functions for the tasks and to make ShellCheck happy.
It won't check for hyprshot-gui version or hyprshot even, will always download every time install.sh is ran. I had added this check, but removed since hyprshot don't have a version command yet. Will add it back if they accept my change there.
I changed the dot desktop file field Name to Name=HyprShot GUI. Probably should change the file name to hyprshot-gui.desktop too, but I'm not sure about it.

Checks for some distros to help installing the required packages (not sure about all distros package names);
Directly downloads hyprshot, hyprshot-gui and the desktop file from the GitHub repositories instead of installing with package managers. Installs hyprshot and hyprshot-gui using the install command and copies;
Looks for the window rules configuration files in case is a modular Hyprland installation or if there are other files (backups etc) inside the hypr folder. Will not change the file if more then one is found. Will copy the configuration to the clipboard with wl-copy to be pasted by the user;
Tries to verify the existance of a hyprshot.conf, creates it with the defaults;